requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
argcomplete==3.5.3
async-timeout==5.0.1
attrs==25.1.0
Babel==2.12.1
beautifulsoup4==4.13.3
boto3==1.28.52
botocore==1.31.85
bs4==0.0.1
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.1.1
cryptography==41.0.4
cssselect==1.2.0
cycler==0.12.1
decorator==5.1.1
dnspython==2.6.1
docopt==0.6.2
docutils==0.15.2
dropbox==9.4.0
email_validator==2.2.0
escpos==1.9
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.8
fiscalyear==0.4.0
fonttools==4.55.8
frozenlist==1.5.0
future==1.0.0
gevent==24.2.1
google-api-core==1.34.1
google-api-python-client==1.8.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
h11==0.14.0
html2text==2020.1.16
htmlmin==0.1.12
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
i18n==0.2
idna==3.10
iniconfig==2.0.0
isodate==0.7.2
Jinja2==3.0.3
jmespath==1.0.1
kiwisolver==1.4.7
libsass==0.22.0
lxml==4.9.3
MarkupSafe==2.1.1
matplotlib==3.6.1
mock==4.0.2
multidict==6.1.0
netifaces==0.11.0
num2words==0.5.12
numpy==1.24.4
Odooly==2.1.9
openai==0.27.4
openpyxl==3.0.7
packaging==24.2
pandas==1.3.5
passlib==1.7.3
pdf2image==1.16.3
pdfkit==1.0.0
phonenumbers==8.13.9
Pillow==9.5.0
platformdirs==4.3.6
pluggy==1.5.0
pretty-bad-protocol==3.1.1
prettytable==2.2.1
propcache==0.2.0
protobuf==3.20.3
psutil==5.9.3
psycopg2==2.9.7
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.20.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.16.1
PyJWT==2.8.0
pyotp==2.6.0
pyparsing==3.1.4
PyPDF2==2.12.1
pypng==0.20220715.0
PyQRCode==1.2.1
pyserial==3.5
pytest==8.3.4
python-barcode==0.15.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.20
python-stdnum==1.18
pytz==2025.1
pyusb==1.2.1
PyYAML==6.0.2
qrcode==7.4.2
redis==5.0.3
reportlab==3.6.13
requests==2.31.0
requests-file==2.1.0
requests-toolbelt==1.0.0
rsa==4.9
s3transfer==0.6.2
simplejson==3.19.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.4.4
starlette==0.45.3
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==3.0.1
urllib3==1.26.20
uvicorn==0.22.0
uWSGI==2.0.28
vatnumber==1.2
viivakoodi==0.8.0
vobject==0.9.6.1
wcwidth==0.2.13
webp==0.1.8
websockets==11.0.3
Werkzeug==2.3.8
xlrd==1.2.0
XlsxWriter==3.1.4
xlwt==1.3.0
yarl==1.15.2
zeep==4.2.1
zope.event==5.0
zope.interface==7.2
zxcvbn==4.4.28
